Added component focus to improved handling of keyboard shortcuts
The way the component was set up originally it would always receive
keyboard shortcuts. This however, caused weird behaviour and errors when
no data was loaded.

So I made it focusable and made the event handler check that either the
component or one of it's children is selected. Also added check to make
sure data is loaded before accepting keyboard input.